chore: interfaces 폴더 안에 modal.ts 생성
interface popup 이동
interface DetailModalProps 이동
